Вынес дебаунсы (поиск в строке и клик на результаты выдачи) из SearchActivity во VM
